remove changes that belong in scalePropertySupportedByScaleType branch
update tsconfig.json


move field back to encoding.ts


fieldDefShorthand and fix import


split enumspec.test from query.test


split enumspec.ts from query.ts


fieldDefShorthand and remove stringify comments


remove additions that belong in scaleProperties PR


add files to cql.ts


fix import for all other files


move relevant test to reflect new query structure and fix import paths


remove old query and query test


extract query methods and correct import paths


scaleType unit test


scalePropertySupportedByScaleType

Note: This currently uses an unreleased dependency from VL. Cannot merge until the dependency is released or I decide to move that code directly into CompassQL
