+++ src/enumerator.ts
@@ -5,7 +5,7 @@ import {checkEncoding} from './constraint/encoding';
 import {checkSpec} from './constraint/spec';
 import {EnumSpecIndex, EnumSpecIndexTuple, SpecQueryModel} from './model';
 import {Property, ENCODING_PROPERTIES, NESTED_ENCODING_PROPERTIES} from './property';
-import {EnumSpec} from './query';
+import {EnumSpec} from './enumspec';
 import {Schema} from './schema';
 
 export let ENUMERATOR_INDEX: {[prop: string]: EnumeratorFactory} = {};
@@ -118,4 +118,3 @@ export function EncodingPropertyGeneratorFactory(prop: Property): EnumeratorFact
     };
   };
 }
-
