+++ src/query/transform.ts
@@ -1,0 +1,12 @@
+import {EnumSpec, ShortEnumSpec} from '../enumspec';
+import {Field} from './encoding';
+
+export interface TransformQuery {
+  filter: FilterQuery[];
+}
+
+export interface FilterQuery {
+  field: Field | EnumSpec<Field> | ShortEnumSpec;
+  operator: string;
+  operand: any | EnumSpec<any> | ShortEnumSpec;
+}
