+++ tsconfig.json
@@ -30,12 +30,17 @@
         "src/cql.ts",
         "src/constraint/encoding.ts",
         "src/constraint/spec.ts",
+        "src/query/encoding.ts",
+        "src/query/shorthand.ts",
+        "src/query/spec.ts",
+        "src/query/transform.ts",
         "src/ranking/effectiveness/effectiveness.ts",
         "src/ranking/effectiveness/channel.ts",
         "src/ranking/effectiveness/mark.ts",
         "src/ranking/ranking.ts",
         "src/config.ts",
         "src/enumerator.ts",
+        "src/enumspec.ts",
         "src/generate.ts",
         "src/model.ts",
         "src/modelgroup.ts",
@@ -45,16 +50,19 @@
         "src/util.ts",
         "test/constraint/encoding.test.ts",
         "test/constraint/spec.test.ts",
+        "test/query/query.test.ts",
+        "test/query/shorthand.test.ts",
+        "test/query/spec.test.ts",
         "test/ranking/effectiveness/effectiveness.test.ts",
         "test/ranking/effectiveness/channel.test.ts",
         "test/ranking/effectiveness/mark.test.ts",
         "test/enumerator.test.ts",
+        "test/enumspec.test.ts",
         "test/generate.test.ts",
         "test/model.test.ts",
         "test/modelgroup.test.ts",
         "test/nest.test.ts",
         "test/property.test.ts",
-        "test/query.test.ts",
         "test/schema.test.ts",
         "typings/chai.d.ts",
         "typings/datalib.d.ts",
